# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SoundMap API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Recording {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  audioUrl: string;
  latitude: number;
  longitude: number;
  /** Human-readable location name (e.g. "São Paulo, Brazil") */
  location: string;
  authorName: string;
  likes: number;
  /** @nullable */
  durationSeconds?: number | null;
  /**
   * Optional photo of the recording location
   * @nullable
   */
  photoUrl?: string | null;
  tags: string[];
  createdAt: string;
}

export interface CreateRecordingBody {
  title: string;
  /** @nullable */
  description?: string | null;
  audioUrl: string;
  latitude: number;
  longitude: number;
  location: string;
  authorName: string;
  /** @nullable */
  durationSeconds?: number | null;
  /** @nullable */
  photoUrl?: string | null;
  tags?: string[];
}

export type RecordingsSummaryTopLocationsItem = {
  location: string;
  count: number;
};

export interface RecordingsSummary {
  totalRecordings: number;
  totalLikes: number;
  topLocations: RecordingsSummaryTopLocationsItem[];
  recentCount: number;
}

export interface Follow {
  id: number;
  followerName: string;
  followingName: string;
  createdAt: string;
}

export interface CreateFollowBody {
  followerName: string;
  followingName: string;
}

export type ListRecordingsParams = {
  /**
   * Filter by latitude (requires lng and radiusKm)
   */
  lat?: number;
  /**
   * Filter by longitude (requires lat and radiusKm)
   */
  lng?: number;
  /**
   * Filter radius in kilometers
   */
  radiusKm?: number;
};

export type GetFeedParams = {
  /**
   * The username of the person viewing the feed
   */
  followerName: string;
};

export type ListFollowsParams = {
  followerName: string;
};