# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SoundMap API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: recordings
    description: Audio recording operations
  - name: follows
    description: Social following operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /recordings:
    get:
      operationId: listRecordings
      tags: [recordings]
      summary: List all recordings
      description: Returns all audio recordings with location data
      parameters:
        - name: lat
          in: query
          required: false
          schema:
            type: number
          description: Filter by latitude (requires lng and radiusKm)
        - name: lng
          in: query
          required: false
          schema:
            type: number
          description: Filter by longitude (requires lat and radiusKm)
        - name: radiusKm
          in: query
          required: false
          schema:
            type: number
          description: Filter radius in kilometers
      responses:
        "200":
          description: List of recordings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recording"
    post:
      operationId: createRecording
      tags: [recordings]
      summary: Create a new recording
      description: Submit a new audio recording with location info
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecordingBody"
      responses:
        "201":
          description: Created recording
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recording"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /recordings/{id}:
    get:
      operationId: getRecording
      tags: [recordings]
      summary: Get a single recording
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recording
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recording"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteRecording
      tags: [recordings]
      summary: Delete a recording
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /recordings/{id}/like:
    post:
      operationId: likeRecording
      tags: [recordings]
      summary: Like a recording
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated recording
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recording"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /recordings/stats/summary:
    get:
      operationId: getRecordingsSummary
      tags: [recordings]
      summary: Get recordings statistics summary
      description: Returns aggregate counts and top locations
      responses:
        "200":
          description: Summary statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordingsSummary"

  /recordings/featured:
    get:
      operationId: listFeaturedRecordings
      tags: [recordings]
      summary: List featured or most-liked recordings
      description: Returns top recordings by likes
      responses:
        "200":
          description: List of featured recordings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recording"

  /recordings/recent:
    get:
      operationId: listRecentRecordings
      tags: [recordings]
      summary: List recently added recordings
      description: Returns the most recently submitted recordings
      responses:
        "200":
          description: List of recent recordings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recording"

  /feed:
    get:
      operationId: getFeed
      tags: [follows]
      summary: Get personalized feed
      description: Returns recordings from people the follower follows
      parameters:
        - name: followerName
          in: query
          required: true
          schema:
            type: string
          description: The username of the person viewing the feed
      responses:
        "200":
          description: Feed recordings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recording"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /follows:
    get:
      operationId: listFollows
      tags: [follows]
      summary: Get who a user follows
      parameters:
        - name: followerName
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of followed names
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Follow"
    post:
      operationId: followUser
      tags: [follows]
      summary: Follow a user by name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFollowBody"
      responses:
        "201":
          description: Created follow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Follow"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /follows/unfollow:
    post:
      operationId: unfollowUser
      tags: [follows]
      summary: Unfollow a user by name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFollowBody"
      responses:
        "204":
          description: Unfollowed
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Recording:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        audioUrl:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        location:
          type: string
          description: Human-readable location name (e.g. "São Paulo, Brazil")
        authorName:
          type: string
        likes:
          type: integer
        durationSeconds:
          type: ["integer", "null"]
        photoUrl:
          type: ["string", "null"]
          description: Optional photo of the recording location
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - audioUrl
        - latitude
        - longitude
        - location
        - authorName
        - likes
        - tags
        - createdAt

    CreateRecordingBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        audioUrl:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        location:
          type: string
        authorName:
          type: string
        durationSeconds:
          type: ["integer", "null"]
        photoUrl:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
      required:
        - title
        - audioUrl
        - latitude
        - longitude
        - location
        - authorName

    RecordingsSummary:
      type: object
      properties:
        totalRecordings:
          type: integer
        totalLikes:
          type: integer
        topLocations:
          type: array
          items:
            type: object
            properties:
              location:
                type: string
              count:
                type: integer
            required:
              - location
              - count
        recentCount:
          type: integer
      required:
        - totalRecordings
        - totalLikes
        - topLocations
        - recentCount

    Follow:
      type: object
      properties:
        id:
          type: integer
        followerName:
          type: string
        followingName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - followerName
        - followingName
        - createdAt

    CreateFollowBody:
      type: object
      properties:
        followerName:
          type: string
        followingName:
          type: string
      required:
        - followerName
        - followingName