# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SoundMap API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all audio recordings with location data
 * @summary List all recordings
 */
export const ListRecordingsQueryParams = zod.object({
  lat: zod.coerce
    .number()
    .optional()
    .describe("Filter by latitude (requires lng and radiusKm)"),
  lng: zod.coerce
    .number()
    .optional()
    .describe("Filter by longitude (requires lat and radiusKm)"),
  radiusKm: zod.coerce
    .number()
    .optional()
    .describe("Filter radius in kilometers"),
});

export const ListRecordingsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  audioUrl: zod.string(),
  latitude: zod.number(),
  longitude: zod.number(),
  location: zod
    .string()
    .describe('Human-readable location name (e.g. \"São Paulo, Brazil\")'),
  authorName: zod.string(),
  likes: zod.number(),
  durationSeconds: zod.number().nullish(),
  photoUrl: zod
    .string()
    .nullish()
    .describe("Optional photo of the recording location"),
  tags: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
});
export const ListRecordingsResponse = zod.array(ListRecordingsResponseItem);

/**
 * Submit a new audio recording with location info
 * @summary Create a new recording
 */
export const CreateRecordingBody = zod.object({
  title: zod.string(),
  description: zod.string().nullish(),
  audioUrl: zod.string(),
  latitude: zod.number(),
  longitude: zod.number(),
  location: zod.string(),
  authorName: zod.string(),
  durationSeconds: zod.number().nullish(),
  photoUrl: zod.string().nullish(),
  tags: zod.array(zod.string()).optional(),
});

/**
 * @summary Get a single recording
 */
export const GetRecordingParams = zod.object({
  id: zod.coerce.number(),
});

export const GetRecordingResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  audioUrl: zod.string(),
  latitude: zod.number(),
  longitude: zod.number(),
  location: zod
    .string()
    .describe('Human-readable location name (e.g. \"São Paulo, Brazil\")'),
  authorName: zod.string(),
  likes: zod.number(),
  durationSeconds: zod.number().nullish(),
  photoUrl: zod
    .string()
    .nullish()
    .describe("Optional photo of the recording location"),
  tags: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a recording
 */
export const DeleteRecordingParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Like a recording
 */
export const LikeRecordingParams = zod.object({
  id: zod.coerce.number(),
});

export const LikeRecordingResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  audioUrl: zod.string(),
  latitude: zod.number(),
  longitude: zod.number(),
  location: zod
    .string()
    .describe('Human-readable location name (e.g. \"São Paulo, Brazil\")'),
  authorName: zod.string(),
  likes: zod.number(),
  durationSeconds: zod.number().nullish(),
  photoUrl: zod
    .string()
    .nullish()
    .describe("Optional photo of the recording location"),
  tags: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
});

/**
 * Returns aggregate counts and top locations
 * @summary Get recordings statistics summary
 */
export const GetRecordingsSummaryResponse = zod.object({
  totalRecordings: zod.number(),
  totalLikes: zod.number(),
  topLocations: zod.array(
    zod.object({
      location: zod.string(),
      count: zod.number(),
    }),
  ),
  recentCount: zod.number(),
});

/**
 * Returns top recordings by likes
 * @summary List featured or most-liked recordings
 */
export const ListFeaturedRecordingsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  audioUrl: zod.string(),
  latitude: zod.number(),
  longitude: zod.number(),
  location: zod
    .string()
    .describe('Human-readable location name (e.g. \"São Paulo, Brazil\")'),
  authorName: zod.string(),
  likes: zod.number(),
  durationSeconds: zod.number().nullish(),
  photoUrl: zod
    .string()
    .nullish()
    .describe("Optional photo of the recording location"),
  tags: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
});
export const ListFeaturedRecordingsResponse = zod.array(
  ListFeaturedRecordingsResponseItem,
);

/**
 * Returns the most recently submitted recordings
 * @summary List recently added recordings
 */
export const ListRecentRecordingsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  audioUrl: zod.string(),
  latitude: zod.number(),
  longitude: zod.number(),
  location: zod
    .string()
    .describe('Human-readable location name (e.g. \"São Paulo, Brazil\")'),
  authorName: zod.string(),
  likes: zod.number(),
  durationSeconds: zod.number().nullish(),
  photoUrl: zod
    .string()
    .nullish()
    .describe("Optional photo of the recording location"),
  tags: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
});
export const ListRecentRecordingsResponse = zod.array(
  ListRecentRecordingsResponseItem,
);

/**
 * Returns recordings from people the follower follows
 * @summary Get personalized feed
 */
export const GetFeedQueryParams = zod.object({
  followerName: zod.coerce
    .string()
    .describe("The username of the person viewing the feed"),
});

export const GetFeedResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  audioUrl: zod.string(),
  latitude: zod.number(),
  longitude: zod.number(),
  location: zod
    .string()
    .describe('Human-readable location name (e.g. \"São Paulo, Brazil\")'),
  authorName: zod.string(),
  likes: zod.number(),
  durationSeconds: zod.number().nullish(),
  photoUrl: zod
    .string()
    .nullish()
    .describe("Optional photo of the recording location"),
  tags: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
});
export const GetFeedResponse = zod.array(GetFeedResponseItem);

/**
 * @summary Get who a user follows
 */
export const ListFollowsQueryParams = zod.object({
  followerName: zod.coerce.string(),
});

export const ListFollowsResponseItem = zod.object({
  id: zod.number(),
  followerName: zod.string(),
  followingName: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListFollowsResponse = zod.array(ListFollowsResponseItem);

/**
 * @summary Follow a user by name
 */
export const FollowUserBody = zod.object({
  followerName: zod.string(),
  followingName: zod.string(),
});

/**
 * @summary Unfollow a user by name
 */
export const UnfollowUserBody = zod.object({
  followerName: zod.string(),
  followingName: zod.string(),
});