# lib/api-zod/src/generated/types/createRecordingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SoundMap API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRecordingBody {
  title: string;
  /** @nullable */
  description?: string | null;
  audioUrl: string;
  latitude: number;
  longitude: number;
  location: string;
  authorName: string;
  /** @nullable */
  durationSeconds?: number | null;
  /** @nullable */
  photoUrl?: string | null;
  tags?: string[];
}