# lib/api-zod/src/generated/types/recording.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SoundMap API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Recording {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  audioUrl: string;
  latitude: number;
  longitude: number;
  /** Human-readable location name (e.g. "São Paulo, Brazil") */
  location: string;
  authorName: string;
  likes: number;
  /** @nullable */
  durationSeconds?: number | null;
  /**
   * Optional photo of the recording location
   * @nullable
   */
  photoUrl?: string | null;
  tags: string[];
  createdAt: Date;
}